+++ asapdiscovery-alchemy/asapdiscovery/alchemy/schema/network.py
@@ -8,8 +8,7 @@ from pydantic import Field
 
 from .atom_mapping import KartografAtomMapper, LomapAtomMapper, PersesAtomMapper
 from .base import _SchemaBase
-from ..utils import check_ligand_series_uniqueness_and_names
-
+from ._util import check_ligand_series_uniqueness_and_names
 
 class _NetworkPlannerMethod(_SchemaBase, abc.ABC):
     """
